ocr/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.1
pillow==10.4.0
pytesseract==0.3.13
python-multipart==0.0.9
